refactored common functions into McRefCommon
