+++ src/McRefCommon.h
@@ -2,6 +2,7 @@
 #define SRC_MCREFCOMMON_H_
 
 #include "DataLayer.h"
+#include "LocusDataLikelihood.h"
 
 #define TRUE 1
 #define FALSE 0
@@ -32,6 +33,21 @@ bool areAlmostEqual(double eventAge, double combAge);
 double relativeDistance(double dbl1, double dbl2);
 
 double requiredRelativePrecision();
+
+
+
+int lca(int pop1, int pop2);
+
+int getPopFather(int popId);
+
+bool isLeafNode(LikelihoodNode *node);
+
+int getLeafNodePop(int nodeId, int gen);
+
+LikelihoodNode *getNode(int nodeId, int gen);
+
+int numNodes();
+
 
 
 /*=== Debug functions ===*/
