+++ src/TauBounds.h
@@ -26,16 +26,4 @@ void printTauBounds(int iteration, FILE *file);
 
 //TODO - move to util
 
-int lca(int pop1, int pop2);
-
-int getPopFather(int popId);
-
-bool isLeafNode(LikelihoodNode *node);
-
-int getLeafNodePop(int nodeId, int gen);
-
-LikelihoodNode *getNode(int nodeId, int gen);
-
-int numNodes();
-
 #endif //G_PHOCS_TAUBOUNDS_H
